LPS-102120 Rename for consistency
